Refactor some attributes name and add getters

+++ app/src/main/java/br/com/porto/isabel/weather/model/Daily.java
@@ -32,5 +32,5 @@ public class Daily {
     private Double windDirection;
 
     @SerializedName("clouds")
-    private Integer clouds;
+    private Double clouds;
 }
